Have only one Tool per file.
