--- src/article/editor/SupplementaryFileTools.js
@@ -1,17 +1,0 @@
-import UploadTool from '../shared/UploadTool'
-
-class SupplementaryFileTool extends UploadTool {
-  get doesAcceptAllFileTypes () {
-    return true
-  }
-
-  get canUploadMultiple () {
-    return false
-  }
-}
-
-export class ReplaceSupplementaryFileTool extends SupplementaryFileTool {
-  getClassNames () {
-    return 'sc-replace-supplementary-file-tool sc-upload-tool sc-tool'
-  }
-}
